requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7
channels==4.0.0
channels-redis==4.1.0
redis==4.6.0
celery==5.3.4
openai==1.3.5
google-generativeai==0.3.2
Pillow==10.1.0
python-decouple==3.8
django-cleanup==6.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
uvicorn==0.24.0
websockets==12.0
whitenoise==6.6.0
django-extensions==3.2.3